jsdoc, move fileExternalDocumentProxy into documents folder
